Extract image slideshow from image table cell
