+++ ui/src/design-system/components/image-carousel/types.ts
@@ -1,0 +1,4 @@
+export enum CarouselTheme {
+  Default = 'default',
+  Light = 'light',
+}
