+++ ui/src/design-system/components/image-carousel/utils.ts
@@ -1,0 +1,12 @@
+import _ from 'lodash'
+import { CSSProperties } from 'react'
+
+export const getImageBoxStyles = (
+  width: string | number = 100
+): CSSProperties => ({
+  width: _.isNumber(width) ? `${width}px` : width,
+})
+
+export const getPlaceholderStyles = (ratio = 1): CSSProperties => ({
+  paddingBottom: `${(1 / ratio) * 100}%`,
+})
