+++ ui/src/design-system/components/table/image-table-cell/image-table-cell.module.scss
@@ -1,91 +1,3 @@
-@import 'design-system/variables/colors.scss';
-@import 'design-system/variables/typography.scss';
-
-/* Basic styles */
-
 .container {
   padding: 12px 16px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.imageBox {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100px;
-  height: 100px;
-  padding: 10px;
-  border-radius: 8px;
-  background-color: $color-neutral-50;
-  border: 1px solid $color-neutral-100;
-
-  &.light {
-    background-color: $color-generic-white;
-  }
-}
-
-.image {
-  max-width: 100%;
-  max-height: 100%;
-}
-
-/* Slideshow styles */
-
-.row {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 10px;
-
-  :not(:last-child) {
-    margin-right: 10px;
-  }
-}
-
-.slide {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 10px;
-  box-sizing: border-box;
-
-  &:not(.visible) {
-    opacity: 0;
-    transition-delay: 0;
-  }
-
-  &.visible {
-    opacity: 1;
-    transition-delay: 0;
-  }
-}
-
-.info {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  @include paragraph-x-small();
-  color: $color-neutral-800;
-  text-align: center;
-
-  :not(:last-child) {
-    margin-right: 6px;
-  }
-}
-
-.control {
-  visibility: hidden;
-
-  &.visible {
-    visibility: visible;
-  }
 }
